Added launch hooks needed by Microsoft authentication
Added a meta field to the interface IUser that implements a new interface called IMeta

IMeta contains information about if the account is a mojang or microsoft account. It also contains information stating if a set account is a demo account (I just thought it could be useful)

I also added checks to components/handler.js to use the new field. If the field is absent, a filler value will be used that will make mclc behave as it did before this change.